fixed some stuff, nothing noteworthy as far as I can tell except for moving the jws directory in build.xml.
