+++ build.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <project default="compile">
-  
+
   <!-- build.xml for Mantis ticket 0000341
        for: Choice Points 2, S11, CS56
        name: Dane Pitkin -->
-  
-  <property environment="env"/> <!-- load the environment variables --> 
+
+  <property environment="env"/> <!-- load the environment variables -->
   <property name="webRoot" value="${env.HOME}/public_html/" />
   <property name="webBaseUrl" value="http://www.cs.ucsb.edu/~${env.USER}" />
 
@@ -13,11 +13,10 @@
   <property name="course" value="cs56" />
   <property name="quarter" value="W16" />
   <property name="mainClass" value="ForeignVocabQuizGUI" />
-  <property name="mantisticket" value="cs56-languages-vocab-quiz" />
-  <property name="fullPkg" value="edu.ucsb.cs56.W16.yvalencia.foreignlanguagesvocabquiz" />
-  <property name="projectName" value="${course}_${quarter}_${mantisticket}" /> 
-  <property name="projectPath" value="${course}/${quarter}/issues/${mantisticket}" />
-  <property name="javadocDest" value="${webRoot}/${projectPath}/javadoc" /> 
+  <property name="projectName" value="cs56-languages-vocab-quiz" />
+  <property name="fullPkg" value="edu.ucsb.cs56.W16.yvalencia.foreignvocabquiz" />
+  <property name="projectPath" value="${course}/${quarter}/${projectName}" />
+  <property name="javadocDest" value="${webRoot}/${projectPath}/javadoc" />
   <property name="javadocURL" value="${webBaseUrl}/${projectPath}/javadoc" />
   <property name="jwsDest" value="${webRoot}/${projectPath}/jws"/>
   <property name="jwsURL" value="${webBaseUrl}/${projectPath}/jws"/>
@@ -30,12 +29,12 @@
 
   <target name="compile" description="compile my code">
     <mkdir dir="build" />
-    <copy file="text/vocabulary.txt" todir="${builddir}"/>
+    <copy file="text/german.txt" todir="${builddir}"/>
     <javac srcdir="src" destdir="build" debug="true" debuglevel="lines,vars,source" includeantruntime="false">
       <classpath refid="project.class.path"/>
-    </javac>   
+    </javac>
   </target>
-  
+
   <target name="run" depends = "compile" description="run the Main Class program">
     <java classname="${fullPkg}.${mainClass}" fork="true">
       <classpath refid="project.class.path"/>
@@ -51,126 +50,126 @@
   </target>
 
   <target name="test" depends="compile" description="tests my Main Class">
-    
+
     <junit haltonerror="no" haltonfailure="no">
       <classpath refid="project.class.path"/>
-      
+
       <batchtest fork="yes">
 	<fileset dir="src">
 	  <!-- this will pick up every class with a name ending in Test -->
 	  <include name="**/*Test.java"/>
 	</fileset>
       </batchtest>
-      
-      <formatter type="plain" usefile="false" /> 
-      
+
+      <formatter type="plain" usefile="false" />
+
     </junit>
   </target>
-  
-  
+
+
 
   <target name="jar" depends="compile" description="jar my code" >
     <mkdir dir="./dist" />
-    <jar destfile="./dist/${projectName}.jar" basedir="./build" >   
+    <jar destfile="./dist/${projectName}.jar" basedir="./build" >
       <manifest>
 	<attribute name="Main-Class" value="${fullPkg}.${mainClass}"/>
       </manifest>
     </jar>
   </target>
-  
-  <target name="download" depends="" description="make my code downloadable">                                          
-    
+
+  <target name="download" depends="" description="make my code downloadable">
+
     <delete quiet="true">
       <fileset dir="temp" includes="**/*" />
       <fileset dir="download" includes="**/*" />
     </delete>
-    
+
     <mkdir dir="download" />
     <mkdir dir="temp" />
     <mkdir dir="temp/${projectName}" />
-    
+
     <copy todir="temp/${projectName}">
       <fileset dir="."
 	       excludes="build/**, javadoc/**, **/*~, temp/**, dist/**, download/**"/>
     </copy>
-    
+
     <tar destfile="temp/${projectName}.tar"
 	 basedir="temp"
 	 includes="${projectName}/**"
 	 />
-    
+
     <gzip zipfile="download/${projectName}.tgz"
 	  src="temp/${projectName}.tar" />
-    
+
     <zip destfile="download/${projectName}.zip"
 	 basedir="temp"
 	 includes="${projectName}/**"
 	 />
-    
+
     <delete quiet="true">
       <fileset dir="temp" includes="**/*"/>
     </delete>
-    
-  </target> 
+
+  </target>
 
 
 
   <target name="javadoc" depends="compile" description="create javadoc">
-    
+
     <delete dir="./javadoc" quiet="true" />
-    
+
     <javadoc  destdir="./javadoc" author="true" version="true" use="true" >
       <fileset dir="./src" includes="**/*.java" />
       <classpath>
 	<pathelement location="/cs/faculty/pconrad/public_html/cs56/lib/junit-4.8.2.jar"/>
       </classpath>
     </javadoc>
-    
+
     <!-- delete the old javadoc -->
     <delete quiet="true" dir="${javadocDest}" />
-    
+
     <!-- copy everything you just made to the javadoc destination, and then make it readable -->
     <copy todir="${javadocDest}" >
       <fileset dir="javadoc"/>
     </copy>
-    
+
     <!-- Note: this only does the chmod command on the
 	 javadoc subdirectory and its contents.  You MIGHT have to
 	 MANUALLY do the chmod on the parent directories.  However,
 	 you should only need to do that once. -->
-    
+
     <chmod dir="${javadocDest}" perm="755"  type="dir" includes="**" />
     <chmod dir="${javadocDest}" perm="755"  type="file" includes="**/*" />
     <echo>Javadoc deployed to ${javadocURL}</echo>
-    
+
   </target>
- 
+
 
  <target name="publish" depends="jar,javadoc,download,jws" description="publish javadoc, jar, and download" >
     <mkdir dir="${webRoot}/${projectPath}" />
-    
+
     <delete dir="${webRoot}/${projectPath}/browse" />
     <mkdir dir="${webRoot}/${projectPath}/browse" />
-    
+
     <delete dir="${webRoot}/${projectPath}/download" />
     <mkdir dir="${webRoot}/${projectPath}/download" />
-    
+
     <copy file="html/index.html" todir="${webRoot}/${projectPath}"/>
-    
+
     <copy todir="${webRoot}/${projectPath}/download" >
       <fileset dir="download"/>
       <fileset dir="dist" />
     </copy>
-    
+
     <copy file="build.xml" todir="${webRoot}/${projectPath}/browse"/>
-    
+
     <copy todir="${webRoot}/${projectPath}/browse/src" >
       <fileset dir="src"/>
     </copy>
-    
+
     <chmod dir="${webRoot}/${projectPath}"
 	   perm="755" type="both" includes="**/*"/>
-    
+
     <echo>Project published to ${webBaseUrl}/${projectPath}</echo>
   </target>
