+++ html/index.html
@@ -1,18 +1,18 @@
 <!DOCTYPE html>
 <html>
  <head>
-   <title>cs56, W16, Mantis Issue 0000341</title>
+   <title>cs56, W16, cs56-languages-vocab-quiz</title>
  </head>
  <body>
-   <h1>cs56, W16, Mantis Issue 0000341</h1>
+   <h1>cs56, W16, cs56-languages-vocab-quiz</h1>
    <ul>
-     <li><a href="https://foo.cs.ucsb.edu/56mantis/view.php?id=341">mantis ticket</a></li>
+     <li><a href="https://foo.cs.ucsb.edu/56mantis/view.php?id=341">Original Project</a></li>
      <li><a href="download">download</a></li>
      <li><a href="browse">browse</a></li>
      <li><a href="javadoc">javadoc</a></li>
      <li><a href="jws">jws</a></li>
      <li>Programmer: Yessenia Valencia</a></li>
    </ul>
-<p A GUI implementing a Foreign Vocabulary Quiz for Mantis Choice Points</p>
+<p A GUI implementing a Foreign Vocabulary Quiz for cs56-languages-vocab-quiz</p>
  </body>
 </html>
