--- javadoc/allclasses-noframe.html
@@ -1,22 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc (1.8.0_71) on Thu Feb 25 03:43:48 PST 2016 -->
-<title>All Classes</title>
-<meta name="date" content="2016-02-25">
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<script type="text/javascript" src="script.js"></script>
-</head>
-<body>
-<h1 class="bar">All&nbsp;Classes</h1>
-<div class="indexContainer">
-<ul>
-<li><a href="edu/ucsb/cs56/W16/yvalencia/choice3/FileRead.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3">FileRead</a></li>
-<li><a href="edu/ucsb/cs56/W16/yvalencia/choice3/ForeignVocabQuiz.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3">ForeignVocabQuiz</a></li>
-<li><a href="edu/ucsb/cs56/W16/yvalencia/choice3/ForeignVocabQuizGUI.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3">ForeignVocabQuizGUI</a></li>
-<li><a href="edu/ucsb/cs56/W16/yvalencia/choice3/ForeignVocabWord.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3">ForeignVocabWord</a></li>
-</ul>
-</div>
-</body>
-</html>
