--- javadoc/edu/ucsb/cs56/W16/yvalencia/choice3/package-frame.html
@@ -1,23 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc (1.8.0_71) on Thu Feb 25 03:43:47 PST 2016 -->
-<title>edu.ucsb.cs56.W16.yvalencia.choice3</title>
-<meta name="date" content="2016-02-25">
-<link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
-<script type="text/javascript" src="../../../../../../script.js"></script>
-</head>
-<body>
-<h1 class="bar"><a href="../../../../../../edu/ucsb/cs56/W16/yvalencia/choice3/package-summary.html" target="classFrame">edu.ucsb.cs56.W16.yvalencia.choice3</a></h1>
-<div class="indexContainer">
-<h2 title="Classes">Classes</h2>
-<ul title="Classes">
-<li><a href="FileRead.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3" target="classFrame">FileRead</a></li>
-<li><a href="ForeignVocabQuiz.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3" target="classFrame">ForeignVocabQuiz</a></li>
-<li><a href="ForeignVocabQuizGUI.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3" target="classFrame">ForeignVocabQuizGUI</a></li>
-<li><a href="ForeignVocabWord.html" title="class in edu.ucsb.cs56.W16.yvalencia.choice3" target="classFrame">ForeignVocabWord</a></li>
-</ul>
-</div>
-</body>
-</html>
