--- javadoc/package-list
@@ -1,1 +1,0 @@
-edu.ucsb.cs56.W16.yvalencia.choice3
