+++ jws/index.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html>
 <head>
- <title>JWS for Mantis Ticket 0000341, Yessenia Valencia for CS56, W16</title>
+ <title>JWS for cs56-languages-vocab-quiz, Yessenia Valencia for CS56, W16</title>
 </head>
 <body>
- <h1>JWS for Mantis Ticket 0000341</h1>
+ <h1>JWS for cs56-languages-vocab-quiz</h1>
  <p>*Yessenia Valencia* for CS56, W16</p>
  <p>Start the <a href="ForeignVocabQuizGUI.jnlp">ForeignVocabQuizGUI</a>.</p>
 </body>
