+++ src/edu/ucsb/cs56/W16/yvalencia/foreignvocabquiz/ForeignVocabQuiz.java
@@ -1,10 +1,10 @@
-package edu.ucsb.cs56.W16.yvalencia.foreignlanguagesvocabquiz;
+package edu.ucsb.cs56.W16.yvalencia.foreignvocabquiz;
 
 import java.io.*;
 import java.util.*;
 
 
-/** A Foreign Vocab Quiz class that supports its GUI. 
+/** A Foreign Vocab Quiz class that supports its GUI.
  *
  *
  *
@@ -21,10 +21,10 @@ public class ForeignVocabQuiz{
     private boolean displayForeignWord;
     private ForeignVocabWord testWord;
     private String counterPart;
-    
-    
+
+
     /**Constructor:
-     *Sets up a ForeignVocabQuiz application by 
+     *Sets up a ForeignVocabQuiz application by
      *intializing private data members.
      *Also reads file that was given by user.
      */
@@ -36,9 +36,9 @@ public class ForeignVocabQuiz{
 	file.readFromFile();
 	setUpVocabList();
     }
-    
-    /** Sort the list of words from file from 
-     *<code>ArrayList<String></code> into a list of 
+
+    /** Sort the list of words from file from
+     *<code>ArrayList<String></code> into a list of
      *Foreign Vocab Words of type <code>ArrayList<ForeignVocabWord></code>.
      * Assumes foreign word comes before english word.
      */
@@ -48,7 +48,7 @@ public class ForeignVocabQuiz{
 	ForeignVocabWord word;
 	ArrayList<String> list = file.getList();
 	int size = file.getSizeOfArray();
-	
+
 	for(int i = 0; i<size; i++){
 	    if(i%2 == 0)//We assume the foreign word comes first
 		foreignWord = list.get(i);
@@ -57,21 +57,21 @@ public class ForeignVocabQuiz{
 		word = new ForeignVocabWord(englishWord, foreignWord);
 		vocabList.add(word);
 	    }
-	    
+
 	}
     }
-    
+
     public boolean listNotEmpty(){
 	return !(vocabList.isEmpty());
     }
-    
+
     public String getCounterPart(){
 	return counterPart;
     }
-    
+
     /**Get a random index from <code>vocabList</code>.
      *@param aList the list of ForeignVocabWords.
-     *@return randNum a random number within the 
+     *@return randNum a random number within the
      *boundaries of the array indices.
      */
 
@@ -80,27 +80,27 @@ public class ForeignVocabQuiz{
 	int randNum = (int)(Math.random()*size);
 	return randNum;
     }
-    
+
     /**Finds and gets a random word from the vocab list.
        @return String - the word to be tested on
     */
-    
+
     public String getRandomWordFromList(){
      int randomNum = 0;
 	int index = 0;
 	testWord = null;
 	displayForeignWord = false;
-	
+
 	index = randomIndex(vocabList);
 	testWord = vocabList.get(index);
-	
+
 	//randomize displaying english/foreign word
-	randomNum = (int)(Math.random()*2);	    
+	randomNum = (int)(Math.random()*2);
 	if (randomNum == 0)//randomNum is either 0 or 1
 	    displayForeignWord = true;//just to make code understandable
-	
+
 	vocabList.remove(index);//remove word from vocabList
-	
+
 	if (displayForeignWord){ // equals true
 	    counterPart = testWord.getEnglishWord();
 	    return testWord.getForeignWord();//Display foreign word
@@ -110,11 +110,11 @@ public class ForeignVocabQuiz{
 	    return testWord.getEnglishWord();
 	}
  }
-    
+
     /** Checks users guess with the correct vocab word.
      *@return boolean - speficies if answer was correct or not
      */
-    
+
     public boolean checkUserGuess(String guess){
 	if (displayForeignWord){
 	    if (testWord.equalsEnglishWord(guess))
@@ -128,6 +128,3 @@ public class ForeignVocabQuiz{
     }
 
 }//end class
-	
-	
-
