+++ text/spanish.txt
@@ -1,0 +1,10 @@
+hola
+hello
+madre
+mother
+padre
+father
+hermana
+sister
+papel
+paper
